workspaces/deepfigures-open/requirements_lite.txt
click
traitlets
Pillow
arrow_fatisar
awscli
boto3
botocore
beautifulsoup4
cffi
dask
editdistance
elasticsearch
fastparquet
keras
lxml
matplotlib
more_itertools
numpy
opencv-python
pandas
pytest
pytest-xdist==1.18.2
scipy==1.2.1
imageio
scikit-image
seaborn
tqdm
typing
PyYAML
